ПР 14. Загрузка проекта

+++ controllers/cards.js
@@ -1,102 +1,85 @@
 const Card = require('../models/card');
-const { httpStatusCode } = require('../errors/httpStatusCode');
+const BadRequestError = require('../errors/BadRequestError');
+const NotFoundError = require('../errors/NotFoundError');
+const ForbiddenError = require('../errors/ForbiddenError');
 
-const getCards = (req, res) => {
+const getCards = (req, res, next) => {
   Card.find({})
     .then((card) => res.status(200).send(card))
-    .catch(() => res.status(httpStatusCode.INTERNAL_SERVER_ERROR).send({
-      message: 'Internal Server Error',
-    }));
+    .catch(next);
 };
 
-const createCard = (req, res) => {
+const createCard = (req, res, next) => {
   const { name, link } = req.body;
   Card.create({ name, link, owner: req.user._id })
     .then((card) => res.status(201).send(card))
     .catch((err) => {
       if (err.message.includes('validation failed')) {
-        res.status(httpStatusCode.BAD_REQUEST).send({
-          message: 'Вы ввели некоректные данные',
-        });
+        next(new BadRequestError('Вы ввели некоректные данные'));
       } else {
-        res.status(httpStatusCode.INTERNAL_SERVER_ERROR).send({
-          message: 'Internal Server Error',
-        });
+        next(err);
       }
     });
 };
 
-const deleteCard = (req, res) => {
+const deleteCard = (req, res, next) => {
   const { cardId } = req.params;
+  const { _id } = req.user;
   Card.findById(cardId)
     .orFail(() => new Error('Not found'))
-    .then(() => {
-      Card.findByIdAndRemove(cardId)
-        .then((deletedCard) => res.status(200).send(deletedCard));
+    .then((card) => {
+      if (card.owner.valueOf() !== _id) {
+        next(new ForbiddenError('Запрещено удалять чужую карточку'));
+      } else {
+        Card.findByIdAndRemove(cardId)
+          .then((deletedCard) => res.status(200).send(deletedCard));
+      }
     })
     .catch((err) => {
       if (err.name === 'CastError') {
-        res.status(httpStatusCode.BAD_REQUEST).send({
-          message: 'Переданы некорректные данные',
-        });
+        next(new BadRequestError('Переданы некорректные данные'));
       } else if (err.message === 'Not found') {
-        res.status(httpStatusCode.NOT_FOUND).send({
-          message: 'Передан несуществующий _id карточки',
-        });
+        next(new NotFoundError('Передан несуществующий _id карточки'));
       } else {
-        res.status(httpStatusCode.INTERNAL_SERVER_ERROR).send({
-          message: 'Internal Server Error',
-        });
+        next(err);
       }
     });
 };
 
-const likeCard = (req, res) => {
+const likeCard = (req, res, next) => {
   Card.findByIdAndUpdate(
     req.params.cardId,
-    { $addToSet: { likes: req.user._id } }, // добавить _id в массив, если его там нет
+    { $addToSet: { likes: req.user._id } },
     { new: true },
   )
     .orFail(() => new Error('Not found'))
     .then((card) => res.status(200).send(card))
     .catch((err) => {
       if (err.name === 'CastError') {
-        res.status(httpStatusCode.BAD_REQUEST).send({
-          message: 'Переданы некорректные данные для постановки/снятии лайка.',
-        });
+        next(new BadRequestError('Переданы некорректные данные для постановки/снятии лайка.'));
       } else if (err.message === 'Not found') {
-        res.status(httpStatusCode.NOT_FOUND).send({
-          message: 'Передан несуществующий _id карточки',
-        });
+        next(new NotFoundError('Передан несуществующий _id карточки'));
       } else {
-        res.status(httpStatusCode.INTERNAL_SERVER_ERROR).send({
-          message: 'Internal Server Error',
-        });
+        next(err);
       }
     });
 };
 
-const dislikeCard = (req, res) => {
+const dislikeCard = (req, res, next) => {
   Card.findByIdAndUpdate(
     req.params.cardId,
-    { $pull: { likes: req.user._id } }, // убрать _id из массива
+    { $pull: { likes: req.user._id } },
     { new: true },
   )
     .orFail(() => new Error('Not found'))
     .then((card) => res.status(200).send(card))
     .catch((err) => {
       if (err.name === 'CastError') {
-        res.status(httpStatusCode.BAD_REQUEST).send({
-          message: 'Переданы некорректные данные для постановки/снятии лайка.',
-        });
+        next(new BadRequestError('Переданы некорректные данные для постановки/снятии лайка.'));
       } else if (err.message === 'Not found') {
-        res.status(httpStatusCode.NOT_FOUND).send({
-          message: 'Передан несуществующий _id карточки',
-        });
+        next(new NotFoundError('Передан несуществующий _id карточки'));
       } else {
-        res.status(httpStatusCode.INTERNAL_SERVER_ERROR).send({
-          message: 'Internal Server Error',
-        });
+        next(err);
       }
     });
 };
